+++ src/components/index.js
@@ -1,0 +1,4 @@
+const toast = require('./toast');
+const output = require('./output');
+
+module.exports = { toast, output };