+++ src/index.js
@@ -7,28 +7,33 @@ let globalState;
 
 const themePath = path.join(__dirname, '../', 'themes', 'index.json');
 
-function activate(context = {}) {
+function activate(context) {
   globalState = context.globalState;
+  const theme = require('./theme')(globalState);
   const version = context.extension.packageJSON.version;
 
   // Don't sync `last-version`. It needs to be machine dependent
   globalState.setKeysForSync(['theme']);
 
+  // globalState.update('theme', undefined);
+
+  // Initialize `theme` on first install.
   if (!globalState.get('theme')) {
-    // Initialize `theme` on first install.
     globalState.update('theme', fs.read(themePath, { json: true }));
   }
 
+  // Installing extension overwrite theme-file. So reload after each install & `deactivate` will do it's job.
   if (!globalState.get('last-version') || globalState.get('last-version') !== version) {
-    // Installing extension overwrite theme-file. So reload after each install & `deactivate` will do it's job.
     globalState.update('last-version', version).then(() => {
       vscode.commands.executeCommand('workbench.action.reloadWindow');
     });
   }
 
-  const disposableOnEdit = vscode.commands.registerCommand('vslook.edit', () => webview(globalState));
+  const disposableOnEdit = vscode.commands.registerCommand('vslook.edit', () => webview(theme));
+  const disposableOnClone = vscode.commands.registerCommand('vslook.clone', () => theme.clone());
+  const disposableOnExport = vscode.commands.registerCommand('vslook.export', () => theme.export());
 
-  context.subscriptions.push(disposableOnEdit);
+  context.subscriptions.push([disposableOnEdit, disposableOnClone, disposableOnExport]);
 }
 
 function deactivate() {