+++ src/types.js
@@ -1,6 +1,6 @@
 module.exports = {
   COLORS: 'workbench.colorCustomizations',
-  TOKEN_COLORS: 'editor.tokenColorCustomizations',
+  TOKENS: 'editor.tokenColorCustomizations',
   PALETTE_TYPE: 'vslook.palette.type',
   PALETTE_COLORS: 'vslook.palette.colors',
   IGNORED_SETTINGS: 'settingsSync.ignoredSettings',